[FIX] refactored import and exportButton to just emit events that parent handles
